requirements.txt
mcp
httpx
python-dotenv
requests
